# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Study System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type LimitErrorKind =
  (typeof LimitErrorKind)[keyof typeof LimitErrorKind];

export const LimitErrorKind = {
  message: "message",
  quiz: "quiz",
  voice: "voice",
} as const;

export interface LimitError {
  error: string;
  code: string;
  kind: LimitErrorKind;
}

export type UserStateLimits = {
  messagesUsed: number;
  messagesLimit: number;
  quizzesUsed: number;
  quizzesLimit: number;
  voiceUsed: number;
  voiceLimit: number;
};

export type UserStateStreak = {
  currentStreak: number;
  bestStreak: number;
  bestScore: number;
  lastActiveDate: string | null;
};

export interface UserState {
  id: number;
  isPremium: boolean;
  premiumUntil: string | null;
  limits: UserStateLimits;
  streak: UserStateStreak;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface SendChatBody {
  messages: ChatMessage[];
  usedVoice?: boolean;
}

export interface ChatReply {
  reply: string;
}

export type UploadResultKind =
  (typeof UploadResultKind)[keyof typeof UploadResultKind];

export const UploadResultKind = {
  image: "image",
} as const;

export interface UploadResult {
  kind: UploadResultKind;
  filename: string;
  excerpt: string;
  summary: string;
}

export type GenerateQuizBodyDifficulty =
  (typeof GenerateQuizBodyDifficulty)[keyof typeof GenerateQuizBodyDifficulty];

export const GenerateQuizBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type GenerateQuizBodyQuestionType =
  (typeof GenerateQuizBodyQuestionType)[keyof typeof GenerateQuizBodyQuestionType];

export const GenerateQuizBodyQuestionType = {
  objective: "objective",
  theory: "theory",
  fill: "fill",
} as const;

export interface GenerateQuizBody {
  subject: string;
  difficulty: GenerateQuizBodyDifficulty;
  questionType: GenerateQuizBodyQuestionType;
  /**
   * @minimum 1
   * @maximum 30
   */
  numQuestions: number;
  /**
   * @minimum 1
   * @maximum 120
   */
  timeMinutes: number;
  instructions?: string;
}

export type QuizQuestionType =
  (typeof QuizQuestionType)[keyof typeof QuizQuestionType];

export const QuizQuestionType = {
  objective: "objective",
  theory: "theory",
  fill: "fill",
} as const;

export interface QuizQuestion {
  id: string;
  prompt: string;
  type: QuizQuestionType;
  options?: string[];
}

export interface Quiz {
  quizId: string;
  subject: string;
  difficulty: string;
  questionType: string;
  timeMinutes: number;
  questions: QuizQuestion[];
}

export interface QuizAnswer {
  questionId: string;
  answer: string;
}

export interface SubmitQuizBody {
  quizId: string;
  subject: string;
  difficulty: string;
  questionType: string;
  questions: QuizQuestion[];
  answers: QuizAnswer[];
}

export interface QuizQuestionResult {
  questionId: string;
  prompt: string;
  userAnswer: string;
  correctAnswer: string;
  isCorrect: boolean;
  explanation: string;
}

export interface QuizResult {
  quizId: string;
  score: number;
  total: number;
  percent: number;
  results: QuizQuestionResult[];
}

export type PaymentInfoPlansItemId =
  (typeof PaymentInfoPlansItemId)[keyof typeof PaymentInfoPlansItemId];

export const PaymentInfoPlansItemId = {
  week: "week",
  month: "month",
} as const;

export type PaymentInfoPlansItem = {
  id: PaymentInfoPlansItemId;
  label: string;
  priceLabel: string;
};

export interface PaymentInfo {
  accountName: string;
  accountNumber: string;
  provider: string;
  plans: PaymentInfoPlansItem[];
}

export interface PaymentSubmission {
  id: number;
  plan: string;
  transactionId: string;
  status: string;
  createdAt: string;
}

export interface AdminLoginBody {
  secretKey: string;
  password: string;
}

export interface AdminToken {
  token: string;
}

export interface AdminUser {
  id: number;
  createdAt: string;
  isPremium: boolean;
  premiumUntil: string | null;
  currentStreak: number;
  bestStreak: number;
  bestScore: number;
  messagesUsedToday: number;
  quizzesUsedToday: number;
}

export interface AdminPayment {
  id: number;
  userId: number;
  plan: string;
  transactionId: string;
  screenshotName: string | null;
  status: string;
  createdAt: string;
}

export type UpgradeUserBodyPlan =
  (typeof UpgradeUserBodyPlan)[keyof typeof UpgradeUserBodyPlan];

export const UpgradeUserBodyPlan = {
  week: "week",
  month: "month",
} as const;

export interface UpgradeUserBody {
  plan: UpgradeUserBodyPlan;
}

export type AdminTokenParameter = string;

export type UploadFileBody = {
  file: Blob;
};

export type SubmitPaymentBodyPlan =
  (typeof SubmitPaymentBodyPlan)[keyof typeof SubmitPaymentBodyPlan];

export const SubmitPaymentBodyPlan = {
  week: "week",
  month: "month",
} as const;

export type SubmitPaymentBody = {
  plan: SubmitPaymentBodyPlan;
  transactionId: string;
  screenshot?: Blob;
};