# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Study System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payment/submit:
    post:
      operationId: submitPayment
      tags: [payment]
      summary: Submit a manual payment proof
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [plan, transactionId]
              properties:
                plan:
                  type: string
                  enum: [week, month]
                transactionId:
                  type: string
                screenshot:
                  type: string
                  format: binary
      responses:
        "200":
          description: Payment submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSubmission"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login (secret key + password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - $ref: "#/components/parameters/AdminToken"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/payments:
    get:
      operationId: adminListPayments
      tags: [admin]
      summary: List all payment submissions (admin)
      parameters:
        - $ref: "#/components/parameters/AdminToken"
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPayment"
  /admin/payments/{id}/approve:
    post:
      operationId: adminApprovePayment
      tags: [admin]
      summary: Approve a payment and upgrade user
      parameters:
        - $ref: "#/components/parameters/AdminToken"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPayment"
  /admin/payments/{id}/reject:
    post:
      operationId: adminRejectPayment
      tags: [admin]
      summary: Reject a payment
      parameters:
        - $ref: "#/components/parameters/AdminToken"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPayment"
  /admin/users/{id}/upgrade:
    post:
      operationId: adminUpgradeUser
      tags: [admin]
      summary: Manually upgrade a user
      parameters:
        - $ref: "#/components/parameters/AdminToken"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradeUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/revoke:
    post:
      operationId: adminRevokeUser
      tags: [admin]
      summary: Revoke a user's premium access
      parameters:
        - $ref: "#/components/parameters/AdminToken"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
components:
  parameters:
    AdminToken:
      name: x-admin-token
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PaymentSubmission:
      type: object
      properties:
        id:
          type: integer
        plan:
          type: string
        transactionId:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required: [id, plan, transactionId, status, createdAt]
    AdminLoginBody:
      type: object
      properties:
        secretKey:
          type: string
        password:
          type: string
      required: [secretKey, password]
    AdminToken:
      type: object
      properties:
        token:
          type: string
      required: [token]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
        isPremium:
          type: boolean
        premiumUntil:
          type: string
          nullable: true
        currentStreak:
          type: integer
        bestStreak:
          type: integer
        bestScore:
          type: integer
        messagesUsedToday:
          type: integer
        quizzesUsedToday:
          type: integer
      required:
        - id
        - createdAt
        - isPremium
        - premiumUntil
        - currentStreak
        - bestStreak
        - bestScore
        - messagesUsedToday
        - quizzesUsedToday
    AdminPayment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        plan:
          type: string
        transactionId:
          type: string
        screenshotName:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - plan
        - transactionId
        - screenshotName
        - status
        - createdAt
    UpgradeUserBody:
      type: object
      properties:
        plan:
          type: string
          enum: [week, month]
      required: [plan]