# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Study System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user state (limits, premium, streak)
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  isPremium: zod.boolean(),
  premiumUntil: zod.string().nullable(),
  limits: zod.object({
    messagesUsed: zod.number(),
    messagesLimit: zod.number(),
    quizzesUsed: zod.number(),
    quizzesLimit: zod.number(),
    voiceUsed: zod.number(),
    voiceLimit: zod.number(),
  }),
  streak: zod.object({
    currentStreak: zod.number(),
    bestStreak: zod.number(),
    bestScore: zod.number(),
    lastActiveDate: zod.string().nullable(),
  }),
});

/**
 * @summary Send a message and receive an AI response
 */
export const SendChatBody = zod.object({
  messages: zod.array(
    zod.object({
      role: zod.enum(["user", "assistant", "system"]),
      content: zod.string(),
    }),
  ),
  usedVoice: zod.boolean().optional(),
});

export const SendChatResponse = zod.object({
  reply: zod.string(),
});

/**
 * @summary Upload an image; returns AI vision analysis
 */
export const UploadFileBody = zod.object({
  file: zod.instanceof(File),
});

export const UploadFileResponse = zod.object({
  kind: zod.enum(["image"]),
  filename: zod.string(),
  excerpt: zod.string(),
  summary: zod.string(),
});

/**
 * @summary Generate a new quiz
 */
export const generateQuizBodyNumQuestionsMax = 30;

export const generateQuizBodyTimeMinutesMax = 120;

export const GenerateQuizBody = zod.object({
  subject: zod.string(),
  difficulty: zod.enum(["easy", "medium", "hard"]),
  questionType: zod.enum(["objective", "theory", "fill"]),
  numQuestions: zod.number().min(1).max(generateQuizBodyNumQuestionsMax),
  timeMinutes: zod.number().min(1).max(generateQuizBodyTimeMinutesMax),
  instructions: zod.string().optional(),
});

export const GenerateQuizResponse = zod.object({
  quizId: zod.string(),
  subject: zod.string(),
  difficulty: zod.string(),
  questionType: zod.string(),
  timeMinutes: zod.number(),
  questions: zod.array(
    zod.object({
      id: zod.string(),
      prompt: zod.string(),
      type: zod.enum(["objective", "theory", "fill"]),
      options: zod.array(zod.string()).optional(),
    }),
  ),
});

/**
 * @summary Submit a finished quiz and get scored results
 */
export const SubmitQuizBody = zod.object({
  quizId: zod.string(),
  subject: zod.string(),
  difficulty: zod.string(),
  questionType: zod.string(),
  questions: zod.array(
    zod.object({
      id: zod.string(),
      prompt: zod.string(),
      type: zod.enum(["objective", "theory", "fill"]),
      options: zod.array(zod.string()).optional(),
    }),
  ),
  answers: zod.array(
    zod.object({
      questionId: zod.string(),
      answer: zod.string(),
    }),
  ),
});

export const SubmitQuizResponse = zod.object({
  quizId: zod.string(),
  score: zod.number(),
  total: zod.number(),
  percent: zod.number(),
  results: zod.array(
    zod.object({
      questionId: zod.string(),
      prompt: zod.string(),
      userAnswer: zod.string(),
      correctAnswer: zod.string(),
      isCorrect: zod.boolean(),
      explanation: zod.string(),
    }),
  ),
});

/**
 * @summary Get OPay payment details and pricing
 */
export const GetPaymentInfoResponse = zod.object({
  accountName: zod.string(),
  accountNumber: zod.string(),
  provider: zod.string(),
  plans: zod.array(
    zod.object({
      id: zod.enum(["week", "month"]),
      label: zod.string(),
      priceLabel: zod.string(),
    }),
  ),
});

/**
 * @summary Submit a manual payment proof
 */
export const SubmitPaymentBody = zod.object({
  plan: zod.enum(["week", "month"]),
  transactionId: zod.string(),
  screenshot: zod.instanceof(File).optional(),
});

export const SubmitPaymentResponse = zod.object({
  id: zod.number(),
  plan: zod.string(),
  transactionId: zod.string(),
  status: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Admin login (secret key + password)
 */
export const AdminLoginBody = zod.object({
  secretKey: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary List all users (admin)
 */
export const AdminListUsersHeader = zod.object({
  "x-admin-token": zod.string(),
});

export const AdminListUsersResponseItem = zod.object({
  id: zod.number(),
  createdAt: zod.string(),
  isPremium: zod.boolean(),
  premiumUntil: zod.string().nullable(),
  currentStreak: zod.number(),
  bestStreak: zod.number(),
  bestScore: zod.number(),
  messagesUsedToday: zod.number(),
  quizzesUsedToday: zod.number(),
});
export const AdminListUsersResponse = zod.array(AdminListUsersResponseItem);

/**
 * @summary List all payment submissions (admin)
 */
export const AdminListPaymentsHeader = zod.object({
  "x-admin-token": zod.string(),
});

export const AdminListPaymentsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  plan: zod.string(),
  transactionId: zod.string(),
  screenshotName: zod.string().nullable(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const AdminListPaymentsResponse = zod.array(
  AdminListPaymentsResponseItem,
);

/**
 * @summary Approve a payment and upgrade user
 */
export const AdminApprovePaymentParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminApprovePaymentHeader = zod.object({
  "x-admin-token": zod.string(),
});

export const AdminApprovePaymentResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  plan: zod.string(),
  transactionId: zod.string(),
  screenshotName: zod.string().nullable(),
  status: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Reject a payment
 */
export const AdminRejectPaymentParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminRejectPaymentHeader = zod.object({
  "x-admin-token": zod.string(),
});

export const AdminRejectPaymentResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  plan: zod.string(),
  transactionId: zod.string(),
  screenshotName: zod.string().nullable(),
  status: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Manually upgrade a user
 */
export const AdminUpgradeUserParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminUpgradeUserHeader = zod.object({
  "x-admin-token": zod.string(),
});

export const AdminUpgradeUserBody = zod.object({
  plan: zod.enum(["week", "month"]),
});

export const AdminUpgradeUserResponse = zod.object({
  id: zod.number(),
  createdAt: zod.string(),
  isPremium: zod.boolean(),
  premiumUntil: zod.string().nullable(),
  currentStreak: zod.number(),
  bestStreak: zod.number(),
  bestScore: zod.number(),
  messagesUsedToday: zod.number(),
  quizzesUsedToday: zod.number(),
});

/**
 * @summary Revoke a user's premium access
 */
export const AdminRevokeUserParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminRevokeUserHeader = zod.object({
  "x-admin-token": zod.string(),
});

export const AdminRevokeUserResponse = zod.object({
  id: zod.number(),
  createdAt: zod.string(),
  isPremium: zod.boolean(),
  premiumUntil: zod.string().nullable(),
  currentStreak: zod.number(),
  bestStreak: zod.number(),
  bestScore: zod.number(),
  messagesUsedToday: zod.number(),
  quizzesUsedToday: zod.number(),
});